feat: add optional imagePull service configuration field
Required to run a service from a locally built container, for which a
"docker pull" command would fail.

This compliments the "image" and "imageTag" service configuration fields
we already have. If the imagePull field is not present, it defaults to
"true". Hence, you only need to set it provide any value when you want
to disable the image being pulled before a deployment or task execution.